backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cassandra-driver>=3.29.3",
    "fastapi>=0.129.0",
    "motor>=3.7.1",
    "neo4j>=6.1.0",
    "pydantic-settings>=2.13.0",
    "python-dotenv>=1.2.1",
    "redis>=7.2.0",
    "uvicorn>=0.41.0",
]
